Refactored ActionState, ActionStateManager
